requirements.txt
Django==4.2.3
djangorestframework==3.14.0
environs==9.5.0
pytest==7.4.0
pytest-django==4.5.2


